# backend/tests/contract/openapi_v1_snapshot.json
{
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "UserLogin": {
    "description": "User login request",
    "properties": {
     "email": {
      "format": "email",
      "title": "Email",
      "type": "string"
     },
     "password": {
      "title": "Password",
      "type": "string"
     }
    },
    "required": [
     "email",
     "password"
    ],
    "title": "UserLogin",
    "type": "object"
   },
   "UserRegister": {
    "description": "User registration request",
    "properties": {
     "email": {
      "format": "email",
      "title": "Email",
      "type": "string"
     },
     "full_name": {
      "title": "Full Name",
      "type": "string"
     },
     "license_number": {
      "title": "License Number",
      "type": "string"
     },
     "municipality": {
      "title": "Municipality",
      "type": "string"
     },
     "password": {
      "title": "Password",
      "type": "string"
     },
     "phone": {
      "title": "Phone",
      "type": "string"
     }
    },
    "required": [
     "email",
     "full_name",
     "phone",
     "password",
     "municipality",
     "license_number"
    ],
    "title": "UserRegister",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "scheme": "bearer",
    "type": "http"
   },
   "OAuth2PasswordBearer": {
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "/api/v1/auth/login"
     }
    },
    "type": "oauth2"
   }
  }
 },
 "info": {
  "contact": {
   "email": "support@valuadis.et",
   "name": "ValuAdis Support",
   "url": "https://valuadis.et/"
  },
  "description": "Ethiopian Property Valuation Platform API with PostGIS spatial support",
  "license": {
   "name": "MIT",
   "url": "https://opensource.org/licenses/MIT"
  },
  "title": "ValuAdis API",
  "version": "1.0.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/auth/login": {
   "post": {
    "description": "Login user and return JWT tokens",
    "operationId": "login_api_v1_auth_login_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UserLogin"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Login",
    "tags": [
     "Authentication",
     "Authentication"
    ]
   }
  },
  "/api/v1/auth/logout": {
   "post": {
    "description": "Logout the browser session by clearing the refresh cookie.\n\nThe presented refresh token (cookie or bearer header) is denylisted so\nit cannot be replayed. Access tokens are short-lived and stateless;\nclients drop them locally.",
    "operationId": "logout_api_v1_auth_logout_post",
    "parameters": [
     {
      "in": "cookie",
      "name": "valuadis_refresh",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Valuadis Refresh"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Logout",
    "tags": [
     "Authentication",
     "Authentication"
    ]
   }
  },
  "/api/v1/auth/me": {
   "get": {
    "description": "Get current user information",
    "operationId": "get_current_user_api_v1_auth_me_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Get Current User",
    "tags": [
     "Authentication",
     "Authentication"
    ]
   }
  },
  "/api/v1/auth/refresh": {
   "post": {
    "description": "Refresh access token \u2014 rotates the refresh token on every call.\n\nAccepts the refresh token as a bearer header (mobile) or as the\nhttpOnly valuadis_refresh cookie (browser). The header wins when both\nare present.",
    "operationId": "refresh_token_api_v1_auth_refresh_post",
    "parameters": [
     {
      "in": "cookie",
      "name": "valuadis_refresh",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Valuadis Refresh"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Refresh Token",
    "tags": [
     "Authentication",
     "Authentication"
    ]
   }
  },
  "/api/v1/auth/register": {
   "post": {
    "description": "Register new user",
    "operationId": "register_api_v1_auth_register_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UserRegister"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Register",
    "tags": [
     "Authentication",
     "Authentication"
    ]
   }
  },
...
 },
...
}